+++ source/extensions/isaacsim.asset.validation/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.1] - 2025-05-30
+### Changed
+- Update to typed name for schema instead of hard-coded strings
+
 ## [1.0.0] - 2025-05-02
 ### Added
 - First Version